A-Gradle: Add Gradle Support
