+++ data/tasks.txt
@@ -1,0 +1,4 @@
+T | 0 | read book
+T | 1 | run 5km
+D | 0 | run 10km | 2021-11-19 | 18:00
+T | 0 | run
