--- src/main/java/Duke.java
@@ -1,61 +1,0 @@
-import command.Command;
-
-import exceptions.DukeException;
-
-import parser.Parser;
-
-import storage.Storage;
-
-import task.TaskList;
-
-import ui.Ui;
-
-public class Duke {
-
-    private final Storage storage;
-    private TaskList tasks;
-    private final Ui ui;
-
-    /**
-     * Constructor for Duke.
-     * Loads the tasks from the file.
-     * If fail, start with an empty TaskList.
-     *
-     * @param filePath File path of the Storage txt file.
-     */
-    public Duke(String filePath) {
-        ui = new Ui();
-        storage = new Storage(filePath);
-        try {
-            tasks = new TaskList(storage.load());
-        } catch (DukeException e) {
-            ui.showError(e.getMessage());
-            tasks = new TaskList();
-        }
-    }
-
-    /**
-     * Runs the Duke program.
-     */
-    public void run() {
-        ui.showWelcome();
-        boolean isExit = false;
-        while (!isExit) {
-            try {
-                String fullCommand = ui.readCommand();
-                ui.showLine(); // show the divider line ("_______")
-                Command c = Parser.parse(fullCommand);
-                c.execute(tasks, ui, storage);
-                isExit = c.isExit();
-            } catch (DukeException e) {
-                ui.showError(e.getMessage());
-            } finally {
-                ui.showLine();
-            }
-        }
-    }
-
-    public static void main(String[] args) {
-        new Duke("data/tasks.txt").run();
-    }
-}
