+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,27 @@
+package duke.command;
+
+import duke.storage.Storage;
+
+import duke.task.TaskList;
+
+import duke.ui.Ui;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showGoodbye();
+    }
+
+    /**
+     * Returns true.
+     * Exit the program.
+     *
+     * @return true
+     */
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+
+}
