+++ text-ui-test/ACTUAL.TXT
